backend/pyproject.toml
[project]
name = "ocr-translate-backend"
version = "1.0.0"
description = "OCR and Translation System with AI-powered corrections"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "httpx>=0.27.2",
    "cryptography>=43.0.1",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "pillow>=11.0.0",
    "markdown>=3.7",
    "reportlab>=4.2.5",
    "numpy>=2.1.3",
    "python-dotenv>=1.0.1",
    "pdf2image==1.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ['py310']
